[HOP-1584] - Verification & Registration Script Changes according to New Flow (New Implementation)
Team - though we restrict ourselves to creating new accounts please make a note of below while we create a new account. The script needs to be changed for verification smoke test.When we create SolePropPass we need to use the below to test properly, basically ensure all SSN are starting 1 because we are doing our own KYB and KYC.

 Also
KYC and KYB would need LastName Specific so choose SolePropPass or BizPass and for BankPartner Routing -> DBAName will need to start with TestTP in lieu of (N) pick a range and stick to it.